SiteReader working again!
Rebuilt the structure of the program. Have a __main__.py that just calls
the CLI. Seems a bit clunky. Will look into what was causing the issue,
why this solution works, and how I can make it a bit tidier. Regardless,
right now it works again, so I'm exhausted and done for the night.